requirements.txt
aiopg==0.7.0
aiozmq==0.7.0
autobahn[asyncio,accelerate]==0.10.5
django==1.8.4
msgpack-python==0.4.6
ujson==1.33
psycopg2==2.6.1
numpy
scipy
shapely==1.5.9
noise==1.2.2
jsonschema==2.5.1
ipython
fake-factory
-e hg+https://dkostochko123@bitbucket.org/schinckel/django-postgres@6a6078d#egg=django-postgres
